Moved the SettingsOverride helper class to tests/helpers.py
