+++ filer/tests/admin.py
@@ -13,8 +13,7 @@ from filer.models.clipboardmodels import Clipboard
 from filer.models.virtualitems import FolderRoot
 from filer.models import tools
 from filer.tests.helpers import (create_superuser, create_folder_structure,
-                                 create_image)
-from filer.test_utils.context_managers import SettingsOverride
+                                 create_image, SettingsOverride)
 from filer.utils.generate_filename import by_path
 
 import filer.settings as filer_settings
